requirements.txt
one-sdk>=1b
python-dotenv==1.0.0
Brotli==1.0.9
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
urllib3==2.0.3
wasmtime==10.0.0
Flask==2.3.2
